backend/requirements.txt
# FastAPI and web server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
psycopg2-binary>=2.9.9  # For synchronous database operations in Celery tasks
alembic>=1.14.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<5.0.0
python-multipart==0.0.12

# Background tasks
celery==5.4.0
redis==5.2.0

# WebSocket
python-socketio>=5.11.4
aiohttp>=3.11.0

# Core ETL dependencies (from old_app)
snowflake-connector-python>=3.6.0
cryptography>=41.0.7

# Data processing
pandas>=1.5.0
numpy>=1.21.0

# Google APIs
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.80.0

# HTTP requests
requests>=2.28.0

# Configuration
pydantic>=2.12.0
pydantic-settings>=2.5.2
email-validator>=2.0.0

# Utilities
python-dotenv==1.0.1

